FINALLY straightened out the whole touch, drag, click wakkey point stuff.
Forever this stuff worked, ok.. But not reliable or terribly accurate. rewrote the event manager stuff. Went over and rethought it all out. Then changed how it was being utilized in drawObj. And now.. It finally works! the flakey scrolling was from getting bogus values from the hardware. Actually it wasn't really bouguis. It had no data and once I guessed that the z point could flag it.. I was able to patch the code cleanly. Yay!